consolidated left side into its own component. Separated right side to show total unacked alerts and a list of all alerts if no nodes are selected, and nodedetails with alerts when a node is selected
